requirements.txt
langchain==0.2.0
langgraph==0.0.50
openai==1.30.1
langchain-openai==0.1.7
pandas
sqlite3
streamlit
python-dotenv==0.19.1